make example more clear with comments
