+++ examples/with-supabase/app/page.tsx
@@ -10,6 +10,8 @@ export default async function Index() {
   const cookieStore = cookies()
 
   const canInitSupabaseClient = () => {
+    // This function is just for the interactive tutorial.
+    // Feel free to remove it once you have Supabase connected.
     try {
       createClient(cookieStore)
       return true
